fix: enforce balance checks and use BigDecimal for all mini-core amounts
- ConversionService: add insufficient-balance check before executing the
  4-leg conversion (was allowing overdrafts like Sarah's -98.10 USD case)
- MiniCoreClient.createTransaction: accept BigDecimal instead of double so
  Jackson serializes at exact declared precision (eliminates 422 from
  mini-core rejecting amounts like 5253.83924303 on 2dp accounts)
- ConversionService: pass fromAmountBD/toAmountBD (truncated BigDecimals)
  instead of raw double fromAmount / double toAmount
- WalletService.p2pTransfer: truncate amount to currency decimal_places
  before sending to mini-core
- UserAccountProvisioningService: use BigDecimal("10000.00") for welcome
  credit instead of double literal
- App.tsx: remove non-functional Activity button from desktop sidebar

+++ api/src/main/java/com/matera/digitaltwin/api/client/MiniCoreClient.java
@@ -7,6 +7,7 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -99,7 +100,7 @@ public class MiniCoreClient {
      * @param description optional; stored in json_payload
      * @return the transaction id, or -1 if the call failed
      */
-    public long createTransaction(long accountId, int transactionCode, double amount,
+    public long createTransaction(long accountId, int transactionCode, BigDecimal amount,
                                   String direction, String description) {
         var body = new HashMap<String, Object>();
         body.put("account_id",        accountId);
